一期

+++ service-goods/service-goods-bean/src/main/java/com/wanmi/sbc/goods/bean/vo/GoodsSyncVO.java
@@ -35,7 +35,7 @@ public class GoodsSyncVO implements Serializable {
      * 供应商类型1博库
      */
     @ApiModelProperty("供应商类型1博库")
-    private Byte goodsSupplierType;
+    private Integer goodsSupplierType;
 
     /**
      * isbn
